Extract path connectors component
